This commit bundles a series of improvements for erasure coding.
Add the ability to rewrite files to be erasure-coded in the mapper instead of the client.
Previously, files needed to be read and rewritten by a client when requesting an encoding for an existing file.
This commit makes the calls asynchronously for the client and saves one read of the file as it is rewritten while being read for the actual encoding.

Add integration tests for the ErasureCodingManager.

Add permission checks for all erasure coding RPCs.

Throw exceptions if a client is calling erasure coding RPCs when erasure coding was disabled.

Throw an exception if trying to append to an erasure-coded file.

Throw an exception if trying to encode an already encoded file.