Make StorageID a UUID HDFS-5115.02
The Storage ID is currently generated from the DataNode's IP+Port+Random components. This scheme will not work when we have separate Storage IDs for each storage directory as there is a possibility of conflicts when an unreliable storage is intermittently offline.
Converting it to a UUID makes collisions very unlikely.